Move bluesound service registration to separate module
